+++ connector/CHANGES.rst
@@ -10,6 +10,8 @@ Changelog
 * Fix: assignment of jobs to workers respect the priority of the jobs (lp:1252681)
 * Pass a new parameter to listeners of 'on_record_create' ( vals:  field values of the new record, e.g {'field_name': field_value, ...})
 * Replace the list of updated fields passed to listeners of 'on_record_write' by a dictionary of updated field values e.g {'field_name': field_value, ...}
+* Add the possibility to use 'Modifiers' functions in the 'direct
+  mappings' (details in the documentation of the Mapper class)
 
 2.0.1 (2013-09-12)
 ~~~~~~~~~~~~~~~~~~